pyproject.toml
[project]
name = "copy-trade-platform"
version = "0.1.0"
description = "Fresh V1 foundation for the Copy Trade platform."
requires-python = ">=3.11"
dependencies = []

[tool.pytest.ini_options]
pythonpath = [
  "apps/api/src",
  "workers/copy_engine/src",
  "packages/domain/src",
  "packages/exchange_adapters/src",
  "packages/shared_events/src"
]
testpaths = [
  "apps/api/tests",
  "workers/copy_engine/tests",
  "packages/domain/tests",
  "packages/exchange_adapters/tests",
  "packages/shared_events/tests"
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
src = [
  "apps/api/src",
  "workers/copy_engine/src",
  "packages/domain/src",
  "packages/exchange_adapters/src",
  "packages/shared_events/src"
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=47.0.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "nats-py>=2.14.0",
    "pydantic>=2.13.3",
    "psycopg[binary]>=3.3.3",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "redis>=7.4.0",
    "ruff>=0.15.12",
    "sqlalchemy>=2.0.49",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.46.0",
]
